requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
Flask-Cors>=3.0
PyPDF2>=3.0.0
Pillow>=9.0
pdf2image>=1.16.0
python-magic>=0.4.27
reportlab>=3.6
Werkzeug>=2.0
eventlet>=0.33.0
gunicorn>=20.0
